Move event listener code into its own task module
