+++ relayer-cli/src/commands/listen.rs
@@ -1,25 +1,10 @@
-use crate::prelude::*;
-
-use futures::future::join_all;
 use std::ops::Deref;
-use tokio::sync::mpsc::{channel, Sender};
 
 use abscissa_core::{
-    application::fatal_error,
-    error::BoxError,
-    tracing::{debug, info},
-    Command, Options, Runnable,
+    application::fatal_error, error::BoxError, tracing::debug, Command, Options, Runnable,
 };
 
-use relayer::config::ChainConfig;
-use relayer::event_handler::*;
-use relayer::event_monitor::*;
-
-use ibc::events::IBCEvent;
-use tendermint::chain::Id as ChainId;
-
-use crate::application::APPLICATION;
-use crate::config::Config;
+use crate::{application::APPLICATION, prelude::*, tasks::event_listener};
 
 #[derive(Command, Debug, Options)]
 pub struct ListenCmd {}
@@ -29,7 +14,7 @@ impl ListenCmd {
         let config = app_config().clone();
 
         debug!("launching 'listen' command");
-        listener_task(&config, false).await
+        event_listener::start(&config, false).await
     }
 }
 
@@ -42,41 +27,4 @@ impl Runnable for ListenCmd {
         })
         .unwrap();
     }
-}
-
-pub async fn listener_task(config: &Config, relay: bool) -> Result<(), BoxError> {
-    let (tx, rx) = channel(100);
-    let mut all_futures = Vec::new();
-    for chain_config in &config.chains {
-        info!(chain.id = % chain_config.id, "spawning event monitor for");
-        let mut event_monitor = init_monitor(chain_config.clone(), tx.clone()).await?;
-        let m_handle = tokio::spawn(async move { event_monitor.run().await });
-        all_futures.push(m_handle);
-    }
-
-    info!("spawning main event handler");
-    let mut event_handler = EventHandler::new(rx, relay);
-    let r_handle = tokio::spawn(async move { event_handler.run().await });
-
-    all_futures.push(r_handle);
-    let _res = join_all(all_futures).await;
-
-    Ok(())
-}
-
-async fn init_monitor(
-    chain_config: ChainConfig,
-    tx: Sender<(ChainId, Vec<IBCEvent>)>,
-) -> Result<EventMonitor, BoxError> {
-    let mut event_monitor =
-        EventMonitor::create(chain_config.id, chain_config.rpc_addr.clone(), tx)
-            .await
-            .map_err(|e| format!("couldn't initialize event monitor: {}", e))?;
-
-    event_monitor
-        .subscribe()
-        .await
-        .map_err(|e| format!("couldn't initialize subscriptions: {}", e))?;
-
-    Ok(event_monitor)
 }
