+++ relayer-cli/src/lib.rs
@@ -22,3 +22,4 @@ pub mod config;
 pub mod error;
 pub mod light;
 pub mod prelude;
+pub mod tasks;
